apiFace/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
compreface-sdk==0.6.0
cryptography==46.0.3
fastapi==0.128.0
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-python==4.12.0.88
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
python-multipart==0.0.21
requests==2.32.5
requests-toolbelt==0.9.1
sse-starlette==3.1.2
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.40.0
